actionsの処理を分割
