+++ .envrc
@@ -3,3 +3,4 @@ use asdf
 PATH_add node_modules/.bin
 
 dotenv .env
+# dotenv .env.test
