--- src/lib/__tests__/actions.test.ts
@@ -1,9 +1,0 @@
-import { describe, expect, test } from '@jest/globals';
-import { getInput } from '@/lib/actions';
-
-describe('getInput', () => {
-  test('github actions以外の環境で実行するとundefinedが返る', () => {
-    const options = {};
-    expect(getInput('path', options)).toBeUndefined();
-  });
-});
